cleaned up verification imports
