+++ grovedb/Cargo.toml
@@ -23,7 +23,7 @@ derive_more = { version = "0.99.18" }
 integer-encoding = { version = "4.0.0", optional = true }
 grovedb-costs = { version = "1.0.0-rc.2", path = "../costs", optional = true }
 nohash-hasher = { version = "0.2.0", optional = true }
-indexmap = { version = "2.2.6", optional = true }
+indexmap = { version = "2.2.6"}
 intmap = { version = "2.0.0", optional = true }
 grovedb-path = { version = "1.0.0-rc.2", path = "../path" }
 grovedbg-types = { path = "../grovedbg-types", optional = true }
@@ -57,7 +57,6 @@ full = [
     "integer-encoding",
     "grovedb-costs",
     "nohash-hasher",
-    "indexmap",
     "intmap"
 ]
 visualize = [
